requirements.txt
aiofiles==24.1.0
asyncpg==0.30.0
celery[redis]
fastapi==0.115.12
pydantic-settings==2.9.1
python-multipart==0.0.20
redis==5.3.0
sqlmodel==0.0.24
urllib3==2.4.0
uvicorn[standard]
